requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-huggingface
langchain-groq
chromadb
sentence-transformers
pypdf
python-dotenv
requests